requirements.txt
opencv-python
mediapipe
pyserial
stable-baselines3
numpy
pyyaml
flask
